pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "feedhog"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "apscheduler>=3.10,<4",
    "fastfeedparser",
    "litellm",
    "pyyaml",
    "python-dateutil",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
]

[tool.setuptools]
packages = []
